Move money to another service

+++ src/main/java/com/lastcivilization/userwriteservice/domain/dto/UserDto.java
@@ -7,5 +7,5 @@ public record UserDto (
         String email,
         Long stats,
         Long equipment,
-        int money
+        Long money
 ){ }
